voyage-embd/requirements.txt
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
pathlib2>=2.3.7
PyYAML>=6.0
